return org result

+++ magda-scala-common/src/main/scala/au/csiro/data61/magda/model/Misc.scala
@@ -88,7 +88,7 @@ package misc {
     homePage: Option[String] = None,
     email: Option[String] = None,
     imageUrl: Option[String] = None,
-    dataSetCount: Option[Int] = None)
+    dataSetCount: Option[Long] = None)
 
   case class Location(
     text: Option[String] = None,
