+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/SearchQueryer.scala
@@ -10,7 +10,7 @@ trait SearchQueryer {
   def search(query: Query, start: Long, limit: Int, facetSize: Int): Future[SearchResult]
   def searchFacets(facetType: FacetType, facetQuery: Option[String], generalQuery: Query, start: Long, limit: Int): Future[FacetSearchResult]
   def searchRegions(query: Option[String], start: Long, limit: Int): Future[RegionSearchResult]
-  def searchOrganisations(query: Option[String], start: Long, limit: Int): Future[OrganisationsSearchResult]
+  def searchOrganisations(query: Option[String], start: Int, limit: Int): Future[OrganisationsSearchResult]
 }
 
 sealed trait SearchStrategy {
